requirements.txt
numpy>=1.26.0,<2.0
pandas~=2.2.2
scipy~=1.13.1
scikit-learn~=1.5.0
matplotlib~=3.8.4
seaborn~=0.13.2
scikit-fuzzy~=0.4.2
PyYAML~=6.0.1
Jinja2~=3.1.4
gymnasium~=0.29.1
stable-baselines3[extra]~=2.3.2
cloudpickle~=3.0.0
torch~=2.3.1
torchvision~=0.18.1
torchaudio~=2.3.1
tensorboard~=2.16.2
wandb~=0.17.0
optuna~=3.6.1
jsonschema~=4.21.1
pyarrow~=14.0.2
